+++ app/src/androidTest/java/nextstep/signup/ConfirmPasswordTest.kt
@@ -11,7 +11,7 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 
-class ConfirmPasswordUITest {
+class ConfirmPasswordTest {
 
     @get:Rule
     val composeTestRule = createComposeRule()
