+++ app/src/androidTest/java/nextstep/signup/EmailTest.kt
@@ -10,7 +10,7 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 
-class EmailUITest {
+class EmailTest {
 
     @get:Rule
     val composeTestRule = createComposeRule()
