+++ app/src/androidTest/java/nextstep/signup/PasswordTest.kt
@@ -10,7 +10,7 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 
-class PasswordUITest {
+class PasswordTest {
 
     @get:Rule
     val composeTestRule = createComposeRule()
